// src/libs/react-grid/reactgrid.d.ts
interface CheckboxCell extends Cell {
    type: 'checkbox';
    checked: boolean;
    checkedText?: string;
    uncheckedText?: string;
}
// ...
interface DateCell extends Cell {
    type: 'date';
    date?: Date;
    format?: Intl.DateTimeFormat;
}
// ...
interface EmailCell extends Cell {
    type: 'email';
    text: string;
    validator?: (text: string) => boolean;
    renderer?: (text: string) => React$1.ReactNode;
}
// ...
interface ChevronCell extends Cell {
    type: 'chevron';
    text: string;
    isExpanded?: boolean;
    hasChildren?: boolean;
    parentId?: Id;
    indent?: number;
}
// ...
interface HeaderCell extends Cell, Span {
    type: 'header';
    text: string;
}
// ...
interface NumberCell extends Cell {
    type: 'number';
    value: number;
    format?: Intl.NumberFormat;
    nanToZero?: boolean;
    hideZero?: boolean;
}
// ...
interface TextCell extends Cell {
    type: 'text';
    text: string;
    placeholder?: string;
    validator?: (text: string) => boolean;
    renderer?: (text: string) => React$1.ReactNode;
}
// ...
interface TimeCell extends Cell {
    type: 'time';
    time?: Date;
    format?: Intl.DateTimeFormat;
}
// ...
interface DropdownCell extends Cell {
    type: 'dropdown';
    selectedValue?: string;
    values: OptionType[];
    isDisabled?: boolean;
    isOpen?: boolean;
    inputValue?: string;
}
// ...
declare type DefaultCellTypes = CheckboxCell | DateCell | EmailCell | ChevronCell | HeaderCell | NumberCell | TextCell | TimeCell | DropdownCell;
// ...
declare type Id = number | string;
// ...
interface CellStyle {
    /** CSS `color` property */
    readonly color?: string;
    /** CSS `background` property */
    readonly background?: string;
    /** CSS `overflow` property */
    readonly overflow?: string;
    /** CSS `padding-left` property */
    readonly paddingLeft?: string;
    /** Object that contains all cell's borders properties */
    readonly border?: {
        readonly left?: BorderProps;
        readonly top?: BorderProps;
        readonly right?: BorderProps;
        readonly bottom?: BorderProps;
    };
}
// ...
interface Cell {
    /** Name of cell type, must be unique */
    type: string;
    /** Marks cell as non editable (by default: `false`) */
    nonEditable?: boolean;
    /** `Id` of group to which this cell belongs to */
    groupId?: Id;
    /** Allowed style properties contained in `CellStyle` interface */
    style?: CellStyle;
    /** Additional CSS classes */
    className?: string;
}
// ...
interface Row<TCell extends Cell = DefaultCellTypes> {
    /** Unique `Id` in all rows array */
    readonly rowId: Id;
    /** Array of `Cell` objects */
    readonly cells: TCell[];
    /** Height of each grid row (in default set to `25` in px) */
    readonly height?: number;
    /**
     * Property that allows row to change is position in grid,
     * default: `false` (row reorder implementation is on the developer's side)
     */
    readonly reorderable?: boolean;
}